WIP: implement most of "describe instead of yaml"
* copy from openResource

Make describe output less hostile

There is no behavior to skip "ManagedFields" in kubectl describe, so we
have to parse and strip it out if we want to display friendly output.

* set version for build

* Revert a lot of custom stuff I dreamed up

Turns out there is already a way to get a Describe behavior out of the
resourceUri, and it handles those "read only" behavior things already!

rearranged some things

test this, (it was the result of resolving several hard merge conflicts)

Signed-off-by: Kingdon Barrett <kingdon@weave.works>