Add a documentation viewer
The node-fetch package provides a fetch method which we can use to
retrieve some docs from a remote source, then display it in an editor
panel as webview.

Add a reference to the GitOps VSCode docs (content to-be-created)

Add a menu entry which triggers the webview for the docs on click, in
the docs tree view.

Signed-off-by: Kingdon Barrett <kingdon@weave.works>