alpha.6 - automatically mark prereleases
Based on parsing the semver, find out if the release is a full release
or prerelease and then mark it up as a prerelease if needed.

This will take the place of the old way, which is manually re-marking
any releases that are pre-releases as pre-releases by hand, as an admin.

Signed-off-by: Kingdon Barrett <kingdon@weave.works>