Enable GitOps Extension to call "set kubeconfig"
The "Set Kubeconfig" action from the Kubernetes Tools extension also
sets the kubeconfig that the Flux extension will use. So let's make it
easy for Flux extension users to do the same action; display their set
contribution point in our UI, so users can call it from our tab.

The effect will not take immediately, and a refresh is required. We can
subscribe to the event that a new Kubeconfig is ready, same when a new
context is chosen. (That will be next!)

Signed-off-by: Kingdon Barrett <kingdon@weave.works>