Fix test
* refreshResourcesTree replaces old refresh methods

UX update: refreshResourcesTreeView is now a unified method that
triggers both Source and Workloads to refresh whenever either circle
arrow is clicked for Sources or Workloads.

This remains distinct from the "refresh clusters" method so that we can
minimize process forking, but avoid showing stale information in most
user scenarios.

Signed-off-by: Kingdon Barrett <kingdon@weave.works>