Allow users to disable running `flux check`
When I tested it locally it looked like we're calling `flux check` a bit
more often than we should. This may have been fixed by now, but the flag
remains so users who are experiencing issues from `flux check` can
disable it entirely.

* allow users to additionally suppress debugging messages

Emits some debugging messages that are meant for the other authors of
the extension to notice (and users won't be as interested in how many
times we were apparently going to call flux check)

Signed-off-by: Kingdon Barrett <kingdon@weave.works>