Allow users to disable running `flux check`
For now, this remains a prerelease feature. We should clean up the debug
information messages before it goes to release. I tested this locally
and it looks like we're calling `flux check` a bit more often than we
should.

I think we should fix that first, then put both this feature and that
fix in the next release 0.25.1

Signed-off-by: Kingdon Barrett <kingdon@weave.works>